rename rewind() to reset(); add new rewind() method; add current() method
In the old code, if iproduct rewould an iterator, then froze it, it could not rebuild its
values from the iterator.  The iterator was reset to its original state, so $prev was undef,
not the last value from the iterator before it was rewound.

To fix this, rewind() is renamed to reset(), as that better indicates what it does. A current()
method is added (iproduct really wants the current value of the iterator), and a
new rewind() method which starts the iterator from the beginning, but does not reset the prev and current
values.