make sure next subs with the same refaddr don't cause overwriting iterator attributes in the registry
the new code doesn't wrap next subs if not necessary, but if a sub is
used twice for different iterators, it will overwrite the attributes
in the register, which can cause havoc.  As an example, see
t/libt/lib/MyTest/Tests/May/Iter.pm, which sets

   next => sub {}

In this case Perl will *not* generate a new closed over coderef for that
sub.  To fix this, there's now Utils::_safe_wrap_next which will wrap
a duplicate coderef in a new sub.