backout code to handle duplicate iterator sub references.
duplicated subroutine iterators indicates that something has gone wrong,
probably that a user didn't correctly set up their closures or we
didn't wrap things correctly.

The original code was put in because the test in t/lib returns sub{}
for the iterator, and it seems that Perl optimizes empty subs,
returning the same reference each time the constructor is
called. (Note that apparently Perl doesn't optimize empty subs across
subroutines, just within a single subroutine).

The previous fix was too broad, in that it always wrapped duplicate
subs.  This fix removes all of that and outputs a more useful
error message.