add filtering for the specific exception value signalling an imported iterator's transition to exhausted state
An exception thrown by an imported iterator need not be due to it
being exhausted.  This commit adds identification of specific
exceptions by one of the following:

  * matching on the stringified exception
  * the exception's class
  * a coderef passed the exception and returning a boolean

By setting the 'on_exception_throw' attribute to
"on_exception_passthrough" the exception emitted by the imported
iterator is also emitted by the wrapper.